Task 1A -Implement GUI for Enqueue, Peek, Count.
Added Enqueue functionality via GUI (button1_Click) to add guests to the queue.

Implemented Count method to display the number of guests in the queue.

Added Peek functionality to show details of the first guest in the queue.

Implemented display of the guest with the most funds .

Integrated basic error handling for empty and uninitialized guest queue.