Task 1A - GUI Create Reverse K Times form and integrate with Form1 menu
Created a new form to handle the reversal of the queue elements by a specified number of times (K). The new form allows users to input the value for K and execute the reversal operation on the guest queue. Integrated this functionality into the main Form1 menu, enabling users to access the reverse feature through the GUI.